Fix WLTO-type-mismatch for some arguments of the subroutine surrosim, in order to pass the check on cran

+++ src/frailtypack.h
@@ -284,7 +284,7 @@ F77_SUB(surrosim) (double *don_simul, double *don_simulS1, int *n_obs,
 		  double *p, double *prop_i, double *gamma, double *alpha,
           int *frailt_base, int *random_generator0, int *aleatoire, 
 		  int *nbre_sim, int *graine, int *nbre_don_non_cons, 
-		  int *param_weibull0, int *thetacopule, int *filtre, 
+		  int *param_weibull0, double *thetacopule, int *filtre, 
           int *filtre2, int *type_joint_simul, int *pfs);
 
 void
